+++ refinedstorage2-platform-common/src/main/java/com/refinedmods/refinedstorage2/platform/common/content/CreativeModeTabItems.java
@@ -4,6 +4,7 @@ import com.refinedmods.refinedstorage2.platform.common.Platform;
 import com.refinedmods.refinedstorage2.platform.common.internal.storage.type.FluidStorageType;
 import com.refinedmods.refinedstorage2.platform.common.internal.storage.type.ItemStorageType;
 import com.refinedmods.refinedstorage2.platform.common.item.ProcessorItem;
+import com.refinedmods.refinedstorage2.platform.common.item.block.ControllerBlockItem;
 
 import java.util.Arrays;
 import java.util.function.Consumer;
@@ -23,7 +24,9 @@ public final class CreativeModeTabItems {
 
     private static void appendBlocks(final Consumer<ItemStack> consumer) {
         final Consumer<ItemLike> itemConsumer = item -> consumer.accept(new ItemStack(item));
-        Items.INSTANCE.getAllControllers().stream().map(Supplier::get).forEach(itemConsumer);
+        Items.INSTANCE.getControllers().stream().map(Supplier::get).forEach(itemConsumer);
+        ControllerBlockItem.getAllAtCapacity().forEach(consumer);
+        Items.INSTANCE.getCreativeControllers().stream().map(Supplier::get).forEach(itemConsumer);
         Items.INSTANCE.getCables().stream().map(Supplier::get).forEach(itemConsumer);
         Items.INSTANCE.getImporters().stream().map(Supplier::get).forEach(itemConsumer);
         Items.INSTANCE.getExporters().stream().map(Supplier::get).forEach(itemConsumer);